DL App/requirements.txt
gradio>=4.44.0,<5
torch
torchvision
pillow>=10.0.0
numpy>=1.24.0
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9
